bug fix: stack sizing issue
the default stack size was not large enough for a task, causing
overflow and eventual system lock up.
this decreases the null processes stack and increases the default
stack size.

since the system stack is abandoned and the heap is used to a new
stack. it is recommended to move allocated stack to heap in the
linker options.